Modify api use a single route for user management
